+++ app/src/store/actions/AlgorithmDialogStoreActions.ts
@@ -3,16 +3,16 @@ import { AlgorithmDialogStoreActionTypes as DialogActions } from 'store/actions/
 import { getAlgorithms } from 'store/actions/AlgorithmsStoreActions';
 import { SnowmanDispatch, SnowmanThunkAction } from 'store/messages';
 import { store } from 'store/store';
-import { MagicNotPossibleId } from 'utils/constants';
+import { MagicNotPossibleId } from 'structs/constants';
+import {
+  SUCCESS_TO_ADD_NEW_ALGORITHM,
+  SUCCESS_UPDATE_ALGORITHM,
+} from 'structs/statusMessages';
 import {
   easyPrimitiveAction,
   easyPrimitiveActionReturn,
 } from 'utils/easyActionsFactory';
 import RequestHandler from 'utils/requestHandler';
-import {
-  SUCCESS_TO_ADD_NEW_ALGORITHM,
-  SUCCESS_UPDATE_ALGORITHM,
-} from 'utils/statusMessages';
 
 export const openAddDialog = (): easyPrimitiveActionReturn =>
   easyPrimitiveAction({
