+++ app/src/structs/constants.ts
