+++ app/src/structs/viewMetaInfoCollection.ts
